Refactor AuthorUtil
Prevent circular dependency
