+++ mysocial/follow/follow_util.py
@@ -1,6 +1,7 @@
 import logging
 
-from authors.models.author import Author, from_author_url_to_local_id
+from authors.models.author import Author
+from authors.util import AuthorUtil
 from follow.models import Follow
 
 logger = logging.getLogger(__name__)
@@ -18,7 +19,7 @@ class FollowUtil:
         Remember to catch errors!
         """
         follower_paths = Follow.objects.values_list('actor', flat=True).filter(target=target.get_url(), has_accepted=True)
-        follow_ids = list(map(lambda f: from_author_url_to_local_id(f), follower_paths))
+        follow_ids = list(map(lambda f: AuthorUtil.from_author_url_to_local_id(f), follower_paths))
         # todo(turnip): support remote author
         return Author.objects.filter(official_id__in=follow_ids)
