+++ mysocial/mysocial/settings/local.py
@@ -1,6 +1,5 @@
 from mysocial.settings.base import *
 from mysocial.settings import base
-from remote_nodes.remote_util import RemoteUtil
 
 DEBUG = True
 
@@ -17,6 +16,3 @@ DATABASES = {
 
 if base.CURRENT_DOMAIN is None:
     base.CURRENT_DOMAIN = '127.0.0.1:8000'
-
-# has to be called last to prevent circular dependency error
-RemoteUtil.setup()
