+++ mysocial/mysocial/urls.py
@@ -14,8 +14,14 @@ Including another URLconf
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
 from django.contrib import admin
-from django.urls import include, path, re_path
+from django.urls import include, path
 from drf_spectacular.views import SpectacularAPIView, SpectacularRedocView, SpectacularSwaggerView
+
+from remote_nodes.remote_util import RemoteUtil
+
+# one time script after settings.py
+# has to be called last to prevent circular dependency error
+RemoteUtil.setup()
 
 urlpatterns = [
     path('', include('authors.urls')),
